move the test from metrics.py to functional/functional_impl.py
